Add check that first element in list is get-user
Because the Fitbit api separates out the user's timezone from the data points, there is a need to retrieve that value in order to establish a correct effective time frame for data points. We have established the convention that the Fitbit mapper takes a list of two JSON response nodes, the first is the response from the get-user-info endpoint and the second is the response from the endpoint of the target type of interest (i.e., BMI or body weight). Previously there was no check as to whether that convention was being followed. We now check that the first element in that list is, in fact, the response node from the get-user-info endpoint.