Check for the correct nodes in input list
There was no verification on the ordering of the input nodes for the Fitbit mapper, which would risk errors being propagated or unexpected results. Now we check that the first node is a get-user response and the second node is a response that contains the appropriate array as provided by the getListNodeName() for each mapper subclass.