Update documentation to easily resume work
The Fitbit refactor currently needs one more mapper - heart rate - and then hooking up to the shim itself. Unfortunately, Fitbit datapoint do not have a timezone, which makes it impossible to guarantee that time frames and time points have the correct offset. Currently the offset is set based on the user's set timezone, which may or may not have been set at the time the data point was created.