Skip body measures that are goals not measured
Withings returns body measures that are both goals set by the user and measurements taken by the user together in the body measures endpoint response. We are mapping measures, not goals, so there is a need to filter out these data points. We short circuit the asDataPoint method for each body measure mapper to return empty if it is a goal instead of a measurement. A helper method is used to accomplish this and the method is conservative in that if there is no value for the "category" field, it will assume that it is a measurement so we don't throw it out. However, we have not seen any datapoint without that field.