Change daily calorie to use start and end times instead of duration
The daily calories burned mapper had used a start_date_time and duration of 1 day to capture that all of the steps were completed over a day. However, this could be ambiguous to upstream consumers in what 1 day means exactly, especially if the start_date_time is translated to a different time zone. In order to provide very precise and specific start and end times, that are unambiguous regardless of time zone adjustments, we changed to using to time stamps, one from the beginning of that date (00:00:00) and one at the end (23:59:59).