Use start of day to start of day for full day datapoints
Previously the mapper would capture an entire day as 00:00:00 - 23:59:59 on that date, however this poses challenges when we don't know the time zone for the start and end, which is the case for Withings, since it implies 23 hours, 59 minutes, and 59 seconds have passed for the data that was captured, which may not be the case. Additionally, you may not have sufficient precision in the end datetime to fully ensure that it captures the end of day. It is simpler to go from beginning of day 1, at midnight, to the beginning of day 2 at midnight to capture that entire. This also has a less strong implication around the length of time and semantically suggests that its from beginning of one day to the beginning of the next, or the entire day inclusive.