Keep sleep start and end datetime in UTC
Previously we had set the timezone for the start and end datetime values to use the timezone provided in the datapoint. However, there is no guarantee that this timezone is the same for the beginning and end of a sleep session, in the rare case of traveling while sleeping. So without having timezone for both start and end, we cannot guarantee data reliability. So it is safer to lose a small amount of data around timezone, but maintain the precise values and semantics for the data.