Use 'Z' as timezone for body measures
Previously body measures used the timezone that was found in the body of the endpoint response, however this was the current timezone of the user in their user profile, so it may have incorrectly attributed a time zone, especially to older datapoints. Although these timepoints were technically correct, they may have been shifted to an incorrect timezone. We were implicitly claiming that we knew what the user's time zone was when serializing datapoints, which could be a mistake. This change sets all of these mappers to use 'Z' as the timezone for the offset in the timestamps. This also allowed us to refactor the WithingsBodyMeasuresDataPointMapper to inherit from the WithingsListDataPointMapper and reuse code from that class, since we no longer had to pass an additional parameter, time zone.