Refactor builder property setting into superclass
For all body measures, the effective time frame and user comments are set the same way, so we created two methods in the parent class that can handle setting these properties in the builder for each measure. Also refactored out the check for goal datapoints, which should be skipped, and datapoints where the user for which the datapoint belongs is uncertain, which should be logged and addressed external to the shim.