Limit requests to desired measures without goals
Previously when a user requests a certain body measure type of data from the shim, we would request all body measures data and then let the mapper handle filtering. We have maintained the filters on the mappers, but now request only the types specified by the users; only retrieving what we need to fulfill the user's request.