Refactor measure and header as separate methods
Previously the datapoint creation methods for each subclass (corresponding to each type of Jawbone data) would require the children to call a method in the parent and pass parameters back to the parent. This created dependencies between the parents and children and required the children to know more about the parent than necessary. We decoupled these and now have separate methods for creating the header and measure, the measure which is abstract and implemented by the children and the header, which has a default implementation and can be overridden or extended by children. These two methods remove the need for the children to call back to the parent before returning their results, again, to the parent.