Use 'Z' as offset instead of user info timezone
Previously we had planned to get the user's time zone and then use that as the offset for data points. However, as we dug deeper into the fitbit api we realized that the user info time zone is not actually used or related to the timestamps of the data from the API. Furthermore, Fitbit does not maintain any timezone information, instead using the device's local time to generate datapoints with no regard for the time zone of the device or user. This means that it is impossible for us to reconstruct the correct instant that datapoints occurred. We have adopted the approach, for the time being, to use the utc offset 'Z' for all timestamps. This provides some consistency, though the data returned should be approached with major caution.