Set step count to use duration time frame
Because fitbit does not provide time zones, it makes more sense for us to only attribute UTC time to one timestamp and then provide a duration of 1 day, which we know is in fact correct. That way we are providing as much correct information as we have access to from fitbit. Step count datapoints now have a start time that is the date contained in the response at beginning of day in UTC, then a duration of 1 day.