Return empty list when no normalized data exists
Previously the shim would return an empty response in the case that there were no datapoints returned by the mapper. This would lead the body property to be set as null instead of as an empty list. This was unclear and would have downstream implications. The shim now returns a complete response, but with an empty list.