Update Withings height and weight endpoint names
The measures type enumeration used "body_weight" and "body_height," however the endpoints used the names "weight" and "height," which caused an exception when retrieving data of those types. We have updated the names of the height and weight endpoints to match the type enumerations. Also updated the README with the new endpoint names.