Update README with types not tested with devices
With Withings and Jawbone, we did not have access to devices that captured a few of the types of data that the APIs provide. In total, we did not have real, device data for 4 types - 1 in Withings and 3 in Jawbone. This means we had to rely on the sample data provided by those third parties to test. We are making this explicit in our documentation and providing a call to action for Shimmer users that may have access to that data to help us test and provide feedback on those specific types.