Suppress JsonMapping exceptions in authentication process
When getData is being called in the context of OAuth authentication, an issue with data mapping implies a failure of the getData request, which is not necessarily tied to the authentication process. We suppress these errors because they may incorrectly suggest that the authentication failed, when it fact it succeeded, but the subsequent getData request failed for some other reason. There is code in subsequent lines that handles checking for the actual success of the OAuth process, so we rely on that to confirm success.