Use auth response type to determine auth success
In processing authentication responses, the authentication response object carries a property indicating whether or not authentication is successful. This property is set to successful if the auth token is set and there are not other exceptions in the process. As this response is returned in the callback endpoint handler, we can then take action on redirecting based on its state.