Fix zone offsets in iHealth mappers
iHealth provides unix second timestamps that are offset to the timezone provided with the data, so the timestamps are not in utc, they are in local time. That means the offset times need to be adjusted based on the provided offset, so we now create offset times in utc, knowing that when the offset seconds are rendered in utc, they will have the correct local date and time, but the wrong offset, so we strip the offset and apply the appropriate offset.