Add extra day to enddate for Jawbone requests
We are inclusive of the last date in our requests for data, however the date times that we use are at beginning of day, meaning that previously we were requesting data until the beginning of the final day. But any activities that occurred on that day would be missed. This change adds an extra day to the enddate of the request so that it requests activities that occurred until the beginning of the next day, which covered the entirety of the final day.