Support Jawbone activity distance with decimal
Jawbone documentation stated that the "meters" value in workouts was of type int, which was supported by all of the test data we had ever seen. However, it was brought to our attention that the value can contain a number with decimal values, which failed to map using an integer-based json mapper. We switched the mapper to use BigDecimal instead of integer as the parsing type, which handles both integer and double data at high precision.