Address inconsistency with auth endpoint response
OAuth1 and OAuth2 shims handled requests to the shimmer authorize endpoint in the case where a user was already authorized for a shim. OAuth2 shims would not initialize another OAuth flow and would return that the shim was already authorized. OAuth1 shims would initiate a new OAuth flow and replace the tokens and auth credentials if completed. This inconsistent behavior was confusing.

This change aligns the OAuth1 response with the OAuth2 response, such that when a request is made to the authorize endpoint for an OAuth1 shim that has already been authorized, it will respond with an indication that the shim is already authorized for the user and not initiate a new OAuth exchange.