Handle positive integer time zones
For the sports endpoint, the API provides time zones in a different way than other endpoints. Instead of providing a string in the form "+0700" they provide a single integer - 0, 1, -4, etc. The ZoneOffset class could parse the negative values, however it could not parse positive integers or timezones with the value 0. The class needs to have a '+' in front of the positive/zero values. This change supports those time zones by adding appending a '+' at the beginning to the timezone representation.