Fix dates for physical activity request from iHealth
The sports.json iHealth endpoint uses the start_time and end_time parameters differently than other endpoints, so we have to add an extra day onto the end of the range so that in encompasses the entire final day until the beginning of the next day. The other endpoints don't require this and instead seem to request the entire day of each parameter.