Removes the numToReturn and numToSkip parameters
The Runkeeper and Jawbone shims previously used low default values for the number of items to be returned that potentially cut out data points that should have been retrieved in fulfilling a request made of Shimmer. This change removes numToReturn and numToSkip, which are legacy parameters for the shimmer API. However, the Runkeeper and Jawbone APIs default to even smaller default limits for the number of items returned (25 and 10 respectively). To address this, we have changed these Shims to set the limit values for the Runkeeper and Jawbone shims to be an arbitrarily high number (100,000) to reduce the likelihood that data points will be lost.

This commit resolves #21.