Repackage application as a runnable WAR
The previous build created a fat JAR that could only be run from the command line. This commit modifies the Gradle configuration to build a WAR that can be run inside an application server. Command-line behaviour is preserved by the Spring Boot Gradle plugin which adds provided dependencies to the WAR.

  Resolves issue #31.